Rearrange base Document classes
