+++ src/allejo/stakx/Document/ContentItem.php
@@ -10,10 +10,10 @@ namespace allejo\stakx\Document;
 use allejo\stakx\Engines\Markdown\MarkdownEngine;
 use allejo\stakx\Engines\PlainTextEngine;
 use allejo\stakx\Engines\RST\RstEngine;
-use allejo\stakx\FrontMatter\Document;
+use allejo\stakx\FrontMatter\FrontMatterDocument;
 use allejo\stakx\Manager\TwigManager;
 
-class ContentItem extends Document implements \JsonSerializable
+class ContentItem extends FrontMatterDocument implements \JsonSerializable, \IteratorAggregate, TwigDocumentInterface
 {
     /**
      * The collection this Content Item belongs to.
@@ -149,5 +149,10 @@ class ContentItem extends Document implements \JsonSerializable
             'permalink' => $this->getPermalink(),
             'redirects' => $this->getRedirects(),
         ));
+    }
+
+    public function getIterator()
+    {
+        return (new \ArrayIterator($this->frontMatter));
     }
 }
