+++ src/allejo/stakx/Manager/TrackingManager.php
@@ -7,7 +7,8 @@
 
 namespace allejo\stakx\Manager;
 
-use allejo\stakx\FrontMatter\Document;
+use allejo\stakx\Document\TwigDocumentInterface;
+use allejo\stakx\FrontMatter\FrontMatterDocument;
 use allejo\stakx\System\FileExplorer;
 use Symfony\Component\Finder\SplFileInfo;
 
@@ -181,15 +182,15 @@ abstract class TrackingManager extends BaseManager
     /**
      * Add a FrontMatterObject based object to the tracker.
      *
-     * @param Document    $trackedItem
-     * @param string      $key
+     * @param TwigDocumentInterface $trackedItem
+     * @param string $key
      * @param string|null $namespace
      */
     protected function addObjectToTracker($trackedItem, $key, $namespace = null)
     {
-        if (!($trackedItem instanceof Document))
+        if (!($trackedItem instanceof FrontMatterDocument) && !($trackedItem instanceof TwigDocumentInterface))
         {
-            throw new \InvalidArgumentException('Only objects can be added to the tracker');
+            throw new \InvalidArgumentException('Only TwigDocumentInterface objects can be added to the tracker');
         }
 
         $this->addArrayToTracker($key, $trackedItem, $trackedItem->getRelativeFilePath(), $namespace);
