+++ tests/allejo/stakx/Test/Document/ContentItemTest.php
@@ -12,7 +12,7 @@ use allejo\stakx\Engines\Markdown\MarkdownEngine;
 use allejo\stakx\Engines\RST\RstEngine;
 use allejo\stakx\Exception\FileAwareException;
 use allejo\stakx\Exception\InvalidSyntaxException;
-use allejo\stakx\FrontMatter\Document;
+use allejo\stakx\FrontMatter\FrontMatterDocument;
 use allejo\stakx\FrontMatter\Exception\YamlVariableUndefinedException;
 use allejo\stakx\Test\PHPUnit_Stakx_TestCase;
 use org\bovigo\vfs\vfsStream;
@@ -45,7 +45,7 @@ class ContentItemTests extends PHPUnit_Stakx_TestCase
         $item = $this->createContentItemWithEmptyFrontMatter();
 
         // The only defined keys should be the ones specially defined
-        $this->assertCount(count(Document::$specialFrontMatterKeys), $item->getFrontMatter());
+        $this->assertCount(count(FrontMatterDocument::$specialFrontMatterKeys), $item->getFrontMatter());
     }
 
     public function testContentItemWithValidFrontMatter()
